add jwtRefreshToken part1

+++ Task Management System API/Helpers/JWT.cs
@@ -5,8 +5,9 @@
         public string Issuer { get; set; }
         public string Audience { get; set; }
         public string SecretKey { get; set; }
-        public int ExpireAfterInMinute { get; set; }
+        public float ExpireAfterInMinute { get; set; }
     }
+
